+++ packages/admin/dashboard/src/components/common/user-link/user-link.tsx
@@ -1,5 +1,6 @@
 import { Avatar, Text } from "@medusajs/ui"
 import { Link } from "react-router-dom"
+import { useUser } from "../../../hooks/api/users"
 
 type UserLinkProps = {
   id: string
@@ -31,4 +32,14 @@ export const UserLink = ({
       </Text>
     </Link>
   )
+}
+
+export const By = ({ id }: { id: string }) => {
+  const { user } = useUser(id) // todo: extend to support customers
+
+  if (!user) {
+    return null
+  }
+
+  return <UserLink {...user} />
 }