+++ packages/core/types/src/http/order/admin/payload.ts
@@ -1,3 +1,18 @@
+export interface AdminUpdateOrder {
+  /**
+   * The order's email.
+   */
+  email?: string
+  /**
+   * The order's shipping address.
+   */
+  shipping_address?: OrderAddress
+  /**
+   * The order's billing address.
+   */
+  billing_address?: OrderAddress
+}
+
 export interface AdminCreateOrderFulfillment {
   /**
    * The items to add to the fulfillment.
@@ -81,4 +96,61 @@ export interface AdminRequestOrderTransfer {
   customer_id: string
   internal_note?: string
   description?: string
+}
+
+export interface OrderAddress {
+  /**
+   * The first name of the address.
+   */
+  first_name?: string
+
+  /**
+   * The last name of the address.
+   */
+  last_name?: string
+
+  /**
+   * The phone number of the address.
+   */
+  phone?: string
+
+  /**
+   * The company of the address.
+   */
+  company?: string
+
+  /**
+   * The first address line of the address.
+   */
+  address_1?: string
+
+  /**
+   * The second address line of the address.
+   */
+  address_2?: string
+
+  /**
+   * The city of the address.
+   */
+  city?: string
+
+  /**
+   * The country code of the address.
+   */
+  country_code?: string
+
+  /**
+   * The province/state of the address.
+   */
+  province?: string
+
+  /**
+   * The postal code of the address.
+   */
+  postal_code?: string
+
+  /**
+   * Holds custom data in key-value pairs.
+   */
+  metadata?: Record<string, unknown> | null
 }