+++ packages/core/utils/src/dml/relations/belongs-to.ts
@@ -3,6 +3,7 @@ import { RelationNullableModifier } from "./nullable"
 
 export class BelongsTo<T> extends BaseRelationship<T> {
   type = "belongsTo" as const
+  declare $foreignKey: true
 
   static isBelongsTo<T>(relationship: any): relationship is BelongsTo<T> {
     return relationship?.type === "belongsTo"
@@ -12,6 +13,6 @@ export class BelongsTo<T> extends BaseRelationship<T> {
    * Apply nullable modifier on the schema
    */
   nullable() {
-    return new RelationNullableModifier<T, BelongsTo<T>>(this)
+    return new RelationNullableModifier<T, BelongsTo<T>, true>(this)
   }
 }