+++ packages/admin/admin-bundler/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @medusajs/admin-bundler
 
+## 2.1.0
+
+### Patch Changes
+
+- Updated dependencies [[`18583ed56725b1faa8b19f7e46e8bc0a59dec33c`](https://github.com/medusajs/medusa/commit/18583ed56725b1faa8b19f7e46e8bc0a59dec33c), [`e8f4f7ea2ba33a747a61a13537e073c82a56d29e`](https://github.com/medusajs/medusa/commit/e8f4f7ea2ba33a747a61a13537e073c82a56d29e), [`a5c8cc992ce9f1dbed6b30048f99fa5bae4af05a`](https://github.com/medusajs/medusa/commit/a5c8cc992ce9f1dbed6b30048f99fa5bae4af05a), [`3d5ca155e303548853ba5ca9e6b3d37ce477823f`](https://github.com/medusajs/medusa/commit/3d5ca155e303548853ba5ca9e6b3d37ce477823f), [`c6f955f0b51a795b4f6a1aeaa1de36205d2d32e5`](https://github.com/medusajs/medusa/commit/c6f955f0b51a795b4f6a1aeaa1de36205d2d32e5)]:
+  - @medusajs/dashboard@2.1.0
+  - @medusajs/admin-shared@2.1.0
+  - @medusajs/admin-vite-plugin@2.1.0
+
 ## 2.0.7
 
 ### Patch Changes