+++ packages/design-system/icons/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @medusajs/icons
 
+## 2.1.0
+
+### Patch Changes
+
+- [#10369](https://github.com/medusajs/medusa/pull/10369) [`94f6265dfc074969f2f141a19f2fc33a6ab96af1`](https://github.com/medusajs/medusa/commit/94f6265dfc074969f2f141a19f2fc33a6ab96af1) Thanks [@kasperkristensen](https://github.com/kasperkristensen)! - feat(ui,icons,ui-preset): Sync with latest changes from Figma, and fix Code components
+
 ## 2.0.7
 
 ## 2.0.6